+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+- removed: `EdgeCurrencyInfo.metaTokens`
+- removed: `EdgeNetworkFee` and `EdgeNetworkFee2` type definitions.
+- removed: Flow type definitions. We only support TypeScript now.
+
 ## 2.43.3 (2026-03-10)
 
 - fixed: Fixed IndexDB cache issue on Android causing failed mixFetch initialization.